Updated existing tests for only-onwer check on adding market to an amm
Earlier tests used to fail on the only onwer on adding market to an amm.
This is because the existing tests didn't initialize the amm or initialized
the amm after creating the market. Amm has been intialized appropriately in
these tests. Only owner check in the amm has been uncommented