format project with new version of prettier.
When we upgraded the repo to v2, we also upgraded the version of
pretty-quick, which itself updated the version of prettier we used. This
caused the formatter to change, so we need to do a project-wide format
so that future PR's aren't confused by format changes as well as code
changes.