add seriesId field to AMM trade subgraph events.
We want to be able to query for a particular account's trade events (BTokenBought, BTokenSold, and WTokenSold) for a particular Series. Prior to this commit there was no way to query by SeriesId. Now, a client can compare the BTokenBought and BTokenSold events to see when there are any instances where a user profited by buying low and selling high.