Add fee capability to bToken buy and sell (#71)
* Add fee capability to bToken buy and sell

* refactor pricing logic from the MinterAmm into a AmmDataProvider contract in order to reduce MinterAmm bytecode size.

We needed to add fee logic to the MinterAmm, but the contract was just a couple bytes under the Spurious Dragon 24576 bytes. In order to make space for the fee logic, in this commit we move all of the view functions pertaining to pricing into a separate AmmDataProvider contract. The MinterAmm interface remains unchanged in order to be backwards comptabile. This reduced the MinterAmm bytecode down to 23025 bytes.

* Add helper logic back in to pass unit tests

* Combine trade fee params into 1 setter

* Address feedback in PR #71

* Ignore vs code settings file

* Refactor so bTokenGetCollateralOut and bTokenGetCollateralIn include the fees in the estimations

Co-authored-by: zareth-san <74431275+zareth-san@users.noreply.github.com>