adding reentrancy protection to all MinterAmm and SeriesController functions which interact with external ERC20 or ERC1155 contracts. (#76)
The September 3 hack depended on function callbacks on ERC1155 tokens allowing the attacker to re-enter the MinterAmm.withdrawCapital function. Here we prevent reentrancy on any external/public MinterAmm and SeriesController function that transfers ERC1155 or ERC20 tokens to prevent this in the future. We include ERC20 tokens, even though they do not have a callback, because they may actually be ERC777 (improved ERC20) tokens which _do_ contain a callback function.